pyproject.toml
[project]
name = "crm_api"
version = "0.1.0"
description = "Carreira Alura Python - Checkpoint 2"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.3",
    "django-cors-headers>=4.6.0",
    "django-filter>=25.2",
    "djangorestframework>=3.17.0",
    "djangorestframework-simplejwt>=5.5.1",
    "djoser>=2.3.3",
    "drf-yasg[validation]>=1.21.15",
    "gunicorn>=25.2.0",
    "jsonschema<4.18.0",
    "markdown>=3.10.2",
    "psycopg[binary]>=3.3.3",
    "python-decouple>=3.8",
    "whitenoise>=6.10.1",
]

[dependency-groups]
dev = [
    "coverage>=7.13.5",
    "pytest>=9.0.2",
    "pytest-cov>=7.1.0",
    "pytest-django>=4.12.0",
]
